busybox-1.0.1-rc1.patch http://busybox.net/lists/busybox/2005-July/014974.html

+++ busybox/archival/libunarchive/archive_xread_all_eof.c
@@ -26,7 +26,7 @@ extern ssize_t archive_xread_all_eof(archive_handle_t *archive_handle, unsigned 
 
 	size = bb_full_read(archive_handle->src_fd, buf, count);
 	if ((size != 0) && (size != count)) {
-		bb_perror_msg_and_die("Short read, read %d of %d", size, count);
+		bb_perror_msg_and_die("Short read, read %ld of %ld", (long)size, (long)count);
 	}
 	return(size);
 }
